ui: Create subscribe integration test sub-module
